refactor: streamline field mask handling in Kubernetes client and reflector
